add examples/tabular and a file that executes it

+++ examples/tabular/tabular.py
@@ -144,8 +144,8 @@ def main():
         x_train, x_eval, y_train, y_eval = train_test_split(x, y, test_size=0.25)
         # Train Dataset
         train_dataset = TensorDataset(
-            torch.from_numpy(x_train.to_numpy()),
-            torch.from_numpy(y_train.to_numpy())
+            torch.from_numpy(x_train.to_numpy()).type(torch.float),
+            torch.from_numpy(y_train.to_numpy()).type(torch.long)
         )
         
         # Test Dataset
